[Omnibus4] OSX build box for dd-agent
This is a building box for our Datadog agent on OSX. It differs slightly
from the other build boxes we have. It's using pretty standard tools
(Vagrant, Chef and Berkshelf...) but it's using them in a rather
uncommon way and this, for two different reasons:
- We rely on VirtualBox to run our OSX VM... and Virtualbox doesn't
  support folder sharing and therefore, there's no way to use Vagrant
  Berkshelf (it creates shared folder between the host and the VM under
  the hood but ssshhh, it's a secret).
- We don't know yet whether the build will be run in a VM or directly on
  the build machine so the provisioning process had to be totally
  independant from Vagrant.

So here's how it goes: the machine (VM or not) is provisionned using
chef-solo... this being said, it requires to install chef-solo along
with a couple of other dependencies for which there's no OSX recipe on
the chef-supermarket.

This first provisioning step is taken care of by the `provision.sh`
shell script. As a matter of fact, Vagrant also uses this one in case
it's used to spawn a virtual build machine. This script takes care of
installing everything we need, `chef-solo` being at the center of it.

That same script then runs chef solo on the dd-agent-osx-build recipe.
It's gonna setup an Omnibus build environment on the machine using the
omnibus::default recipe on the chef market. It also takes care of
creating the folders required by omnibus with proper rights on it. Our
recipe is ridiculously small but it actually makes a lot of things under
the hood (all of it having been written by the Chef team).

The the script will pull the dd-agent-omnibus project, generate binstubs
for it and that all. You just have to launch
`dd-agent-omnibus/omnibus_build.sh`, sit back and watch (for quite a
long time actually, 1h30 on a VM on our Macbook pros and 45-50minutes on
real hardware (16GB RAM + 3GHz i7).

Have fun !